Add unit-test for glycan related functions. rearrage some unit tests from NGlyco and OGlyco Test.
