+++ Test/XLTestNOGlyco.cs
@@ -1,0 +1,28 @@
+﻿using Chemistry;
+using EngineLayer;
+using EngineLayer.CrosslinkSearch;
+using EngineLayer.Indexing;
+using MassSpectrometry;
+using NUnit.Framework;
+using Proteomics;
+using Proteomics.Fragmentation;
+using Proteomics.ProteolyticDigestion;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using TaskLayer;
+using UsefulProteomicsDatabases;
+using MzLibUtil;
+using Nett;
+using EngineLayer.GlycoSearch;
+using MathNet.Numerics.LinearRegression;
+
+namespace Test
+{
+    [TestFixture]
+    public class XLTestNOGlyco
+    {
+
+    }
+}
